requirements.txt
appdirs>=1.4.3
CacheControl>=0.12.6
certifi>=2020.6.20
chardet>=3.0.4
colorama>=0.4.3
contextlib2>=0.6.0
cycler>=0.10.0
distlib>=0.3.0
distro>=1.4.0
html5lib>=1.0.1
idna>=2.8
ipaddr>=2.2.0
kiwisolver>=1.2.0
lockfile>=0.12.2
matplotlib>=3.3.1
msgpack>=0.6.2
numpy>=1.19.1
numpy-stl>=2.11.2
packaging>=20.3
pep517>=0.8.2
Pillow>=7.2.0
progress>=1.5
PyOpenGL>=3.1.5
pyparsing>=2.4.7
PyQt5>=5.15.0
PyQt5-sip>=12.8.0
pyqtgraph>=0.11.0
pyserial>=3.4
python-dateutil>=2.8.1
python-utils>=2.4.0
pytoml>=0.1.21
pytz>=2020.1
requests>=2.22.0
retrying>=1.3.3
scipy>=1.5.2
six>=1.15.0
urllib3>=1.25.8
webencodings>=0.5.1
